% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpit_pois.R
\name{dpit_pois}
\alias{dpit_pois}
\title{Residuals for regression models with poisson outcomes}
\usage{
dpit_pois(y, mu)
}
\arguments{
\item{y}{An observed outcome vector.}

\item{mu}{A vector of fitted mean values.}
}
\value{
A \code{dpit} object containing DPIT residuals.
}
\description{
Computes DPIT residuals for Poisson outcomes regression using the observed counts (\code{y}) and their
corresponding fitted mean values (\code{mu}).
}
\details{
For formulation details on discrete outcomes, see \code{\link{dpit}}.
}
\examples{
## Poisson example
n <- 500
set.seed(1234)
# Covariates
x1 <- rnorm(n)
x2 <- rbinom(n, 1, 0.7)
# Coefficients
beta0 <- -2
beta1 <- 2
beta2 <- 1
lambda1 <- exp(beta0 + beta1 * x1 + beta2 * x2)
y <- rpois(n, lambda1)

# True model
poismodel <- glm(y ~ x1 + x2, family = poisson(link = "log"))
y1 <- poismodel$y
p1f <- fitted(poismodel)
dpit.poi <- dpit_pois(y=y1, mu=p1f)
resid.poi <- residuals(dpit.poi)
plot(dpit.poi)

}
