% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpit_znb.R
\name{dpit_znb}
\alias{dpit_znb}
\title{Residuals for regression models with zero-inflated negative binomial outcomes}
\usage{
dpit_znb(y, mu, pzero, size)
}
\arguments{
\item{y}{An observed outcome vector.}

\item{mu}{A vector of fitted mean values for the count (non-zero) component.}

\item{pzero}{A vector of fitted probabilities for the zero-inflation component.}

\item{size}{A dispersion parameter of the negative binomial distribution.}
}
\value{
A \code{dpit} object containing DPIT residuals.
}
\description{
Computes DPIT residuals for regression models with zero-inflated negative
binomial outcomes using the observed counts (\code{y}) and their fitted distributional
parameters (\code{mu}, \code{pzero}, \code{size}).
}
\details{
For formulation details on discrete outcomes, see \code{\link{dpit}}.
}
\examples{
## Zero-Inflated Negative Binomial
library(pscl)
n <- 500
set.seed(1234)

# Covariates
x1 <- rnorm(n)
x2 <- rbinom(n, 1, 0.7)

# Coefficients
beta0 <- -2
beta1 <-  2
beta2 <-  1
beta00 <- -2
beta10 <-  2

# NB dispersion (size = theta; larger => closer to Poisson)
theta_true <- 1.2

# Mean of NB count part
mu_true <- exp(beta0 + beta1 * x1 + beta2 * x2)

# Excess zero probability (logit)
p0 <- 1 / (1 + exp(-(beta00 + beta10 * x1)))

## simulate outcomes
z  <- rbinom(n, size = 1, prob = 1 - p0)              # 1 => from NB, 0 => structural zero
y1 <- rnbinom(n, size = theta_true, mu = mu_true)     # NB count draw
y  <- ifelse(z == 0, 0, y1)

## True model
modelzero1 <- zeroinfl(y ~ x1 + x2 | x1, dist = "negbin", link = "logit")
y1 <- modelzero1$y
mu1    <- stats::predict(modelzero1, type = "count")
pzero1 <- stats::predict(modelzero1, type = "zero")
theta1 <- modelzero1$theta
dpit.zero1 <- dpit_znb(y = y1, pzero = pzero1, mu = mu1, size = theta1)
resid.zero1 <- residuals(dpit.zero1)
plot(dpit.zero1)

## Ignoring zero-inflation: NB only
modelzero2 <- MASS::glm.nb(y ~ x1 + x2)
y2 <- modelzero2$y
mu2    <- fitted(modelzero2)
theta2 <- modelzero2$theta
dpit.zero2 <- dpit_nb(y = y2, mu = mu2, size = theta2)
resid.zero2 <- residuals(dpit.zero2)
plot(dpit.zero2)
}
