% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGPIF.R
\docType{data}
\name{LGPIF}
\alias{LGPIF}
\title{LGPIF Data}
\format{
A data frame with 5677 rows and 41 variables:
\describe{
\item{PolicyNum}{Policy number}
\item{Year}{Policy year}
\item{ClaimBC}{Total building and contents (BC) claims in the year}
\item{ClaimIM}{Total inland marine (IM) claims in the year (contractor’s equipment)}
\item{ClaimPN}{Total comprehensive claims from new motor vehicles in the year}
\item{ClaimPO}{Total comprehensive claims from old motor vehicles in the year}
\item{ClaimCN}{Total collision claims from new vehicles in the year}
\item{ClaimCO}{Total collision claims from old vehicles in the year}
\item{TypeCity}{Indicator for city entity}
\item{TypeCounty}{Indicator for county entity}
\item{TypeMisc}{Indicator for miscellaneous entity}
\item{TypeSchool}{Indicator for school entity}
\item{TypeTown}{Indicator for town entity}
\item{TypeVillage}{Indicator for village entity}
\item{IsRC}{Indicator for replacement cost (motor vehicles)}
\item{CoverageBC}{Log coverage amount for building and contents (in millions of dollars)}
\item{lnDeductBC}{Log deductible amount for building and contents}
\item{NoClaimCreditBC}{Indicator for no BC claims in prior year}
\item{yAvgBC}{Average BC claim amount}
\item{FreqBC}{Frequency of BC claims}
\item{CoverageIM}{Log coverage amount for inland marine (in millions of dollars)}
\item{lnDeductIM}{Log deductible amount for inland marine}
\item{NoClaimCreditIM}{Indicator for no IM claims in prior year}
\item{yAvgIM}{Average IM claim amount}
\item{FreqIM}{Frequency of IM claims}
\item{CoveragePN}{Log coverage amount for comprehensive new vehicles (in millions of dollars)}
\item{NoClaimCreditPN}{Indicator for no PN claims in prior year}
\item{yAvgPN}{Average PN claim amount}
\item{FreqPN}{Frequency of PN claims}
\item{CoveragePO}{Log coverage amount for comprehensive old vehicles (in millions of dollars)}
\item{NoClaimCreditPO}{Indicator for no PO claims in prior year}
\item{yAvgPO}{Average PO claim amount}
\item{FreqPO}{Frequency of PO claims}
\item{CoverageCN}{Log coverage amount for collision of new vehicles (in millions of dollars)}
\item{NoClaimCreditCN}{Indicator for no CN claims in prior year}
\item{yAvgCN}{Average CN claim amount}
\item{FreqCN}{Frequency of CN claims}
\item{CoverageCO}{Log coverage amount for collision of old vehicles (in millions of dollars)}
\item{NoClaimCreditCO}{Indicator for no CO claims in prior year}
\item{yAvgCO}{Average CO claim amount}
\item{FreqCO}{Frequency of CO claims}
}
}
\source{
https://sites.google.com/a/wisc.edu/jed-frees/
}
\usage{
LGPIF
}
\description{
Data from the Wisconsin Local Government Property Insurance Fund (LGPIF).
The LGPIF was established to provide property insurance for local
government entities that include counties, cities, towns,
villages, school districts, fire departments, and other
miscellaneous entities, and is administered by the Wisconsin Office of the Insurance Commissioner. Properties covered under this fund include government buildings, vehicles, and equipment.
}
\references{
Frees, E. W., Lee, G., & Yang, L. (2016). "Multivariate frequency-severity regression models in insurance." \emph{Risks}, 4(1), 4.
}
\keyword{datasets}
