% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2pm.R
\name{resid_2pm}
\alias{resid_2pm}
\title{Residuals for regression models with two-parts outcomes}
\usage{
resid_2pm(model0, model1, y, part0, part1, plot=TRUE, scale = "normal")
}
\arguments{
\item{model0}{glm model for 0 outcomes}

\item{model1}{model for continuous outcoms}

\item{y}{outcome variables}

\item{part0}{cdf for y=0}

\item{part1}{cdf for y>0}

\item{plot}{A logical value indicating whether or not to return QQ-plot}

\item{scale}{You can choose the scale of the residuals among \code{normal} and \code{uniform} scales. The default scale is \code{normal}.}
}
\value{
The double probability integral transform residuals(DPIT residuals).
}
\description{
Calculates DPIT residuals for semi-continuous outcome regression such as 2PM.
}
