% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi.R
\name{resid_quasi}
\alias{resid_quasi}
\title{Quasi Emprical residuals functions}
\usage{
resid_quasi(model)
}
\arguments{
\item{model}{Model object (e.g., \code{glm}, \code{glm.nb}, \code{polr}, \code{zeroinfl})}
}
\value{
A QQ plot.
\itemize{
\item x-axis: Theoretical quantiles
\item y-axis: Sample quantiles
}
}
\description{
Draw the QQ-plot for regression models with discrete outcomes using quasi-empirical residual distribution functions.
Specifically, the model assumption of GLMs with binary, ordinal, Poisson, negative binomial,
zero-inlated Poisson, and zero-inflated negative binomial outcomes can be applicable to \code{resid_quasi()}.
}
\details{
detail here
}
\references{
Lu Yang (2021). Assessment of Regression Models with Discrete Outcomes Using Quasi-Empirical Residual Distribution Functions, Journal of Computational and Graphical Statistics, 30(4), 1019-1035.
}
