% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_resid.R
\name{resid_zeroinfl}
\alias{resid_zeroinfl}
\title{Residuals for zeroinflated regression model}
\usage{
resid_zeroinfl(model, plot=TRUE)
}
\arguments{
\item{model}{glm model object (eg. \code{zeroinfl()} from \code{pscl})}

\item{plot}{A logical value indicating whether or not to return QQ-plot}
}
\value{
proposed residuals
}
\description{
\code{resid_zeroinfl} is used for calculating residuals based on probability integral transform.
A zeroinflated model from \code{pscl} is recommended in this package.
}
\examples{
## zeroinflated poisson model
library(pscl)
# simulation
n <- 1e3
x1 <- rnorm(n)
x2 <- rbinom(n, 1, 0.7)

b1 <- 2;b2 <- 1;b0 <- -1
b00 <- -2;b10 <- 2
lambda1 <- exp(b0 + b1 * x1 + b2 * x2)
p0 <- 1 / (1 + exp(-(b00 + b10*x1)))
y0 <- rbinom(n, size = 1, prob = 1 - p0)
y1 <- rpois(n, lambda1)
y <- ifelse(y0 == 0, 0, y1)

m1 <- zeroinfl(y ~ x1 + x2 | x1, dist = "poisson", link = "logit")
zpois.resid <- resid_zeroinfl(m1) ## with qqplot
zpois.resid <- resid_zeroinfl(m1, plot=FALSE) ## without qqplot
}
