% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ordered_curve.R
\name{ord_curve}
\alias{ord_curve}
\title{Ordered Curve}
\usage{
ord_curve(model, thr)
}
\arguments{
\item{model}{glm model object (eg. \code{glm}, \code{glm.nb}, \code{polr})}

\item{thr}{threshold variable (predictor, fitted values, or variable to be included as a covariate)}
}
\description{
\code{ord.curve} is used to plot the ordered-curve plot.
}
\details{
If the mean structure is well specified in the model,
\eqn{(L_1(t), L_2(t))} should be close to each other.
If the curve is distant from the diagonal, it suggests incorrectness in the mean structure.
Moreover, if the curve is above the diagonal, the summation of the response is larger than
the fitted mean, which implies that the mean is underestimated, and vice versa. \cr

The role of \code{thr}(threshold variable) is to determine the rule for summing each mean and outcome observation.
The candidate for \code{thr} could be any function of predictors such as a single predictor(eg. \code{x1}),
a linear combination of predictor(eg.\code{x1+x2}), or fitted values(eg. \code{fitted(model)}). \cr

For more details, see the reference paper.
}
\examples{
library(glmnet)
n <- 1e2
b <- c(2, -1, -1)
x1 <- rnorm(n); x2 <- rnorm(n)
y <-  rpois(n, exp(b[1]+b[2]*x1+b[3]*x2))

m1 <- glm(y~x1+x2, family=poisson)

ord_curve(m1, x1) # Ordered curve with a threshold, x1
ord_curve(m1, x2) # Ordered curve with a threshold, x2
ord_curve(m1, fitted(m1)) # Ordered curve with a threshold, fitted values

}
