% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie.R
\name{resid_semiconti}
\alias{resid_semiconti}
\title{Residuals for semicontinuous outcome regression}
\usage{
resid_semiconti(model, plot=TRUE)
}
\arguments{
\item{model}{model object(using tweedie family)}

\item{plot}{A logical value indicating whether or not to return QQ-plot}
}
\value{
residuals
}
\description{
\code{resid.semiconti} is used to calculate residuals for semi-continuous outcomes regression such as tweedie model.
A model object of semicontinuous regression from \code{tweedie} package is recommended.
}
\examples{
library(tweedie)
library(statmod)
n <- 1e3
x11<-rnorm(n)
x12 <- rnorm(n)
b1 <- 1; b2<- 1;b0 <- 5
lambda1<-exp(b0+b1*x11+b2*x12)
z <- runif(n,0,1)
y1 <- qtweedie(z, mu=lambda1, xi=1.6,phi=10)
out=tweedie.profile(y1~x11+x12,p.vec=seq(1.1,1.9,length=9),
                    method="interpolation",do.ci=TRUE,do.smooth=TRUE,do.plot=TRUE)
m1 <- glm(y1~x11+x12,family=tweedie(var.power=out$p.max,link.power=0))
tweedie.reisd <- resid_semiconti(m1) # with qqplot
tweedie.reisd <- resid_semiconti(m1, plot=FALSE) #without qqpplot
}
